file changes to make logic more sound
