+++ backend/expense_categorization/merchantCategories.js
@@ -94,4 +94,5 @@ async function createYelpCategory(merchantName, categoryId, confidence) {
 module.exports = {
   categorizeTransaction,
   getOrCreateCategory,
+  createYelpCategory,
 };
